修改buffer读写限制

+++ service/buffer.go
@@ -138,8 +138,8 @@ func newBuffer(size int64) (*buffer, error) {
 		cwait:          0,
 		pwait:          0,
 	}
-	for i := 0; i < size; i++ {
-		_buffer.buf[i] = &ByteArray{index: -1, bArray: make([]byte, 1)}
+	for i := int64(0); i < size; i++ {
+		_buffer.buf[i] = &ByteArray{index: -1, bArray: make([]byte, int64(1))}
 	}
 	return &_buffer, nil
 }
